requirements.txt
pbr>=1.6 # Apache-2.0

aiohttp
PyYAML
